+++ src/main/java/com/gmail/scanner/model/SourceResult.java
@@ -1,7 +1,8 @@
 package com.gmail.scanner.model;
 
 import com.gmail.scanner.service.model.Source;
+import java.math.BigDecimal;
 
-public record SourceResult(Source source, int orders, double sum) {
+public record SourceResult(Source source, int orders, BigDecimal sum) {
 
 }
